move persistence to separate file
